% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_regression_example}
\alias{linear_regression_example}
\title{RcppNumericalSolver example}
\usage{
linear_regression_example()
}
\value{
Numeric vector
}
\description{
An example demonstrating RcppNumericalSolver
}
