% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonnegativeleastsquares_example}
\alias{nonnegativeleastsquares_example}
\title{Non-negative least squares example}
\usage{
nonnegativeleastsquares_example()
}
\value{
Numeric vector
}
\description{
An example demonstrating non-negative least squares with the RcppNumericalSolver library
}
