% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rosenbrock_example}
\alias{rosenbrock_example}
\title{Rosenbrock example}
\usage{
rosenbrock_example()
}
\value{
Numeric vector
}
\description{
An example demonstrating Rosenbrock with the RcppNumericalSolver library.
}
